requirements.txt
yfinance
flask
pillow
beautifulsoup4
googlesearch-python
requests